Add filteredChildren container to CollectionView
Fixes #3580

Besides this bug, the ability to understand what views are currently not filtered out has been a semi-regular request.

This does add an extra container, but I don't think it should cost that much.  I intend to run this against the js-framework-benchmark to see how it affects thing, but I don't think it'll be bad.

The big question I _think_ is in the naming..  I'm not sure `filteredChildren` makes sense as a public API?  In context it kinda does..  _however_ one alternate option is to name _this_ container to `children` and make the current `children` private.  I don't _really_ know a use-case for needing both attached and detached children.

I could see arguments either way.